requirements.txt
pypdf==4.3.1
pdf2docx==0.5.8
PyMuPDF==1.24.9
Pillow==10.4.0

# Packaging
pyinstaller==6.10.0



